moved everything in app into a unified folder system.
