--- next-client/src/app/(home)/examples/page.tsx
@@ -1,15 +1,0 @@
-export default function ExamplesPage() {
-  return (
-    <>
-      <h2>Examples of reports:</h2>
-      <ul>
-        <li>
-          <a href="/examples/heal-michigan">Heal Michigan</a>
-        </li>
-        <li>
-          <a href="/examples/CIP">CIP data</a>
-        </li>
-      </ul>
-    </>
-  );
-}
