+++ next-client/src/app/examples/cip-data/page.tsx
