+++ next-client/src/components/elements/sonner/Sonner.tsx
@@ -2,7 +2,7 @@
 
 import { useTheme } from "next-themes";
 import { Toaster as Sonner } from "sonner";
-import "../../../app/global.css";
+// import "../../../app/global.css";
 
 type ToasterProps = React.ComponentProps<typeof Sonner>;
